% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag.1.R
\name{Diag.1}
\alias{Diag.1}
\title{Diagnostic ggplots for models}
\usage{
Diag.1(My.Model)
}
\arguments{
\item{My.Model}{A model in a list format. eg lm or glm}
}
\value{
A ggplot object; a diagnostic plot (modeled on the base version)
}
\description{
Quick Model Diagnostics in GGplot.
Standardised Residules v Fitted Values
Using the Base version diagnostic plots as a template.
}
\examples{

Test models

lm.1 <- lm(mpg ~ wt,data=mtcars)
glm.1 <- glm(formula= vs ~ wt + disp, data=mtcars, family=binomial)

Example as function
Diag.1(lm.1)
Diag.1(glm.1)

Example with magrittr pipe
lm.1 \%>\% Diag.1()
glm.1 \%>\% Diag.1()

Output shoud be a ggplot object in form of diagnostic plot.

Coresponding Base example
plot(lm.1, which=1)
plot(glm.1, which=1)





}
\author{
Justin Castagna, \email{justin.castagna@students.mq.edu.au}
}
\keyword{diagnostic}
\keyword{ggplot2}
\keyword{model}
