% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag.2.R
\name{Diag.2}
\alias{Diag.2}
\title{Diagnostic ggplots for models}
\usage{
Diag.2(My.Model)
}
\arguments{
\item{My.Model}{A model in a list format. eg lm or glm}
}
\value{
A ggplot object; a diagnostic plot (modeled on the base version)
}
\description{
Quick Model Diagnostics in GGplot.
Using the Base version diagnostic plots as a template.
}
\examples{

Test models

lm.1 <- lm(mpg ~ wt,data=mtcars)
glm.1 <- glm(formula= vs ~ wt + disp, data=mtcars, family=binomial)

Example as function()
Diag.2(lm.1)
Diag.2(glm.1)

Example with magrittr pipe
lm.1 \%>\% Diag.2()
glm.1 \%>\% Diag.2()

Output shoud be a ggplot object in form of diagnostic plot.



}
\author{
Justin Castagna, \email{justin.castagna@students.mq.edu.au}
}
\keyword{diagnostic}
\keyword{ggplot2}
\keyword{model}
