% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag.4.R
\name{Diag.4}
\alias{Diag.4}
\title{Cooks Distance - Diagnostic ggplot for models}
\usage{
Diag.4(My.Model)
}
\arguments{
\item{My.Model}{A model in a list format. eg lm or glm}
}
\value{
A ggplot object; a diagnostic plot (modeled on the base version)
}
\description{
Quick Model Diagnostics in GGplot.
Cooks Distance
Using the Base version diagnostic plots as a template.
}
\examples{

# Cooks Distance
# Test models

lm.1 <- lm(mpg ~ wt,data=mtcars)
glm.1 <- glm(formula= vs ~ wt + disp, data=mtcars, family=binomial)

# Example as function()
Diag.4(lm.1)
Diag.4(glm.1)

# Example with magrittr pipe
lm.1 \%>\% Diag.4()
glm.1 \%>\% Diag.4()

# Output shoud be a ggplot object in form of diagnostic plot.

# Coresponding Base example
plot(lm.1, which=4)
plot(glm.1, which=4)


}
\author{
Justin Castagna, \email{justin.castagna@students.mq.edu.au}
}
\keyword{diagnostic}
\keyword{ggplot2}
\keyword{model}
