% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag.1.R
\name{Diag.1}
\alias{Diag.1}
\title{Diagnostic Test Functions}
\usage{
Diag.1(My.Model)
}
\arguments{
\item{set}{A model list eg lm or glm}
}
\value{
A ggplot diagnostic plot (modeled on the base version)
}
\description{
Question: How to ensure libs are loaded; library(tidyverse), library(broom)
}
\details{
Quick Model Diagnostics in GGplot.
Using the Base version diagnostic plots as a template.
}
\examples{

Test models
lm.1=lm(mpg ~ wt,data=mtcars)
glm.1 <- glm(formula= vs ~ wt + disp, data=mtcars, family=binomial)
Diag.1(lm.1)
Diag.1(glm.1)
lm.1 \%>\% Diag.1
glm.1 \%>\% Diag.1

Output shoud be a ggplot object



}
\author{
Justin, \email{justin.castagna@students.mq.edu.au}
}
\keyword{diagnostic}
\keyword{ggplot2}
\keyword{model}
